--- app/src/main/java/com/example/myassets/presentation/ui/MainActivity.kt
@@ -1,22 +1,0 @@
-package com.example.myassets.presentation.ui
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import androidx.navigation.fragment.NavHostFragment
-import androidx.navigation.ui.setupWithNavController
-import com.example.myassets.databinding.ActivityMainBinding
-import dagger.hilt.android.AndroidEntryPoint
-
-@AndroidEntryPoint
-class MainActivity : AppCompatActivity() {
-    private lateinit var binding: ActivityMainBinding
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        binding = ActivityMainBinding.inflate(layoutInflater)
-        setContentView(binding.root)
-        val navHostFragment = binding.fragmentHolder.getFragment() as NavHostFragment
-        val navController = navHostFragment.navController
-
-        binding.navigationBar.setupWithNavController(navController)
-    }
-}
