--- app/src/main/java/com/example/myassets/presentation/ui/portfolios/rv/listeners/PortfolioListItemListener.kt
@@ -1,6 +1,0 @@
-package com.example.myassets.presentation.ui.portfolios.rv.listeners
-
-interface PortfolioListItemListener {
-    fun onItemClick(itemId: Int)
-    fun onSelect(itemId: Int)
-}
