+++ app/src/main/java/presentation/ui/assets/AssetListFragment.kt
@@ -1,0 +1,31 @@
+package presentation.ui.assets
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.viewModels
+import com.example.myassets.databinding.FragmentAssetListBinding
+import presentation.ui.assets.rv.AssetListRvAdapter
+import presentation.util.BasicFragment
+import dagger.hilt.android.AndroidEntryPoint
+
+@AndroidEntryPoint
+class AssetListFragment : BasicFragment<FragmentAssetListBinding>() {
+    private val viewModel: AssetListViewModel by viewModels()
+
+    override fun inflateViewBinding(
+        inflater: LayoutInflater,
+        container: ViewGroup?
+    ) = FragmentAssetListBinding.inflate(inflater, container, false)
+
+    override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
+        super.onViewCreated(view, savedInstanceState)
+        val adapter = AssetListRvAdapter()
+        binding.assetListRv.adapter = adapter
+
+        viewModel.portfolioItemsList.observe(viewLifecycleOwner) { portfolioItemsList ->
+            adapter.submit(portfolioItemsList)
+        }
+    }
+}
