+++ app/src/main/java/presentation/ui/portfolios/rv/listeners/PortfolioListItemListener.kt
@@ -1,0 +1,6 @@
+package presentation.ui.portfolios.rv.listeners
+
+interface PortfolioListItemListener {
+    fun onItemClick(itemId: Int)
+    fun onSelect(itemId: Int)
+}
