+++ app/src/main/java/presentation/util/BasicFragment.kt
@@ -1,0 +1,28 @@
+package presentation.util
+
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.Fragment
+import androidx.viewbinding.ViewBinding
+
+abstract class BasicFragment<VB : ViewBinding> : Fragment() {
+    private var _binding: VB? = null
+    protected val binding get() = _binding!!
+
+    protected abstract fun inflateViewBinding(inflater: LayoutInflater, container: ViewGroup?): VB
+    override fun onCreateView(
+        inflater: LayoutInflater,
+        container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        _binding = inflateViewBinding(inflater, container)
+        return binding.root
+    }
+
+    override fun onDestroyView() {
+        _binding = null
+        super.onDestroyView()
+    }
+}
