+++ app/src/main/res/layout/fragment_settings.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".presentation.ui.settings.SettingsFragment">
+    tools:context="presentation.ui.settings.SettingsFragment">
 
     <TextView
         android:layout_width="wrap_content"
