+++ app/src/main/res/navigation/main_graph.xml
@@ -6,17 +6,17 @@
     app:startDestination="@id/fragment_home">
     <fragment
         android:id="@+id/fragment_home"
-        android:name="com.example.myassets.presentation.ui.home.HomeFragment"
+        android:name="presentation.ui.home.HomeFragment"
         android:label="homepage"
         tools:layout="@layout/fragment_home"/>
     <fragment
         android:id="@+id/asset_list_fragment"
-        android:name="com.example.myassets.presentation.ui.assets.AssetListFragment"
+        android:name="presentation.ui.assets.AssetListFragment"
         android:label="assetslist"
         tools:layout="@layout/fragment_asset_list" />
     <fragment
         android:id="@+id/portfolio_list_fragment"
-        android:name="com.example.myassets.presentation.ui.portfolios.PortfolioListFragment"
+        android:name="presentation.ui.portfolios.PortfolioListFragment"
         android:label="portfolio"
         tools:layout="@layout/fragment_portfolio_list" >
         <action
@@ -26,7 +26,7 @@
     </fragment>
     <fragment
         android:id="@+id/portfolio_details_fragment"
-        android:name="com.example.myassets.presentation.ui.portfolios.PortfolioDetailsFragment"
+        android:name="presentation.ui.portfolios.PortfolioDetailsFragment"
         android:label="portfolioDetails"
         tools:layout="@layout/fragment_portfolio_details" >
         <argument
@@ -36,7 +36,7 @@
     </fragment>
     <fragment
         android:id="@+id/fragment_settings"
-        android:name="com.example.myassets.presentation.ui.settings.SettingsFragment"
+        android:name="presentation.ui.settings.SettingsFragment"
         android:label="settings"
         tools:layout="@layout/fragment_settings">
         <action
@@ -46,6 +46,6 @@
 
     <dialog
         android:id="@+id/chooseDefaultCurrencyFragment"
-        android:name="com.example.myassets.presentation.ui.settings.stringselector.CurrencyBottomSheetFragment"
+        android:name="presentation.ui.settings.stringselector.CurrencyBottomSheetFragment"
         android:label="ChooseDefaultCurrencyFragment" />
 </navigation>
